Terragrunt added to handle backend and remove duplication
A tfvars file in the root holds the terragrunt configuration
that will handle backend creation & usage, removing this from
modules.

This file also specifies common tfvars variables for each module.

The pipeline module specifies the cloudbuild pipeline for the
project and needs to be created to bootstrap the project. At this
point the pipeline will build and apply the environments/ module.

The structure for the infrastructure is TBC and environments is
simply a placeholder.