Pipeline logic moved into separate module
This pipeline defines the stages of our
application related pipelines. It shouldn't
matter whether this is the front end, elastic
cluster, API gateway for the UI, we should
have the same environment pattern. At this point
this is:
- staging
- production

The builds are unique to each actual pipeline that is
build, so at this stage they need to be passed as a var
to the pipeline.